+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-...
+### Fixed
+- Disallow setting a different slave address through `SlaveAddr::Default`.
 
 ## 0.1.0 - 2018-08-18